Refactor frontend from Angular to React
