+++ nin/frontend/app/index.html
@@ -2,16 +2,11 @@
 <html>
   <head>
     <meta charset=utf-8>
-    <title></title>
+    <title>nin</title>
   </head>
-  <body ng-app=nin>
+  <body>
 
-    <div
-      class=container-fluid
-      keybinding
-      ng-include="'views/main.html'"
-      ng-controller="MainCtrl as main">
-    </div>
+    <div class=container-fluid id="root"></div>
 
     <script src="//localhost:9000/gen/dasBoot.js"></script>
     <script src="//localhost:9000/gen/shaders.js"></script>
