+++ nin/frontend/app/scripts/components/editor/Connection.js
