+++ nin/frontend/app/scripts/components/editor/FlyaroundController.js
