+++ nin/frontend/app/scripts/components/editor/GraphEditorNode.js
