--- nin/frontend/app/scripts/directives/ngRightClick.js
@@ -1,15 +1,0 @@
-/* Inspired by http://stackoverflow.com/a/15732476 */
-
-function ngRightClick($parse) {
-  return function(scope, element, attrs) {
-    var fn = $parse(attrs.ngRightclick);
-    element.bind('contextmenu', function(event) {
-      scope.$apply(function() {
-        event.preventDefault();
-        fn(scope, {$event: event});
-      });
-    });
-  };
-}
-
-module.exports = ngRightClick;
