--- nin/frontend/app/scripts/directives/scrollcallback.js
@@ -1,12 +1,0 @@
-function scrollcallback(demo) {
-  return {
-    restrict: 'A',
-    link: function(scope, element, attrs) {
-      element.on('scroll', function(event) {
-        scope[attrs.scrollcallback](event);
-      });
-    }
-  };
-}
-
-module.exports = scrollcallback;
