--- nin/frontend/app/scripts/directives/selectOnClick.js
@@ -1,14 +1,0 @@
-function selectOnClick($window) {
-  return {
-    restrict: 'A',
-    link: function (scope, element) {
-      element.on('click', function () {
-        if (!$window.getSelection().toString()) {
-          this.setSelectionRange(0, this.value.length);
-        }
-      });
-    }
-  };
-}
-
-module.exports = selectOnClick;
