--- nin/frontend/app/scripts/filters/index.js
@@ -1,9 +1,0 @@
-const angular = require('angular');
-
-const errorExcerpt = require('./errorExcerpt');
-const stackFormat = require('./stackFormat');
-
-module.exports = angular.module('nin.filters', [])
-                   .filter('errorExcerpt', errorExcerpt)
-                   .filter('stackFormat', stackFormat)
-                   .name;
