--- nin/frontend/app/scripts/services/ScriptReloader.js
@@ -1,7 +1,0 @@
-class ScriptReloader {
-  reload(path, callback) {
-    $.getScript(path, callback);
-  }
-}
-
-module.exports = ScriptReloader;
