--- nin/frontend/app/scripts/services/ninrc.js
@@ -1,9 +1,0 @@
-function ninrc($http) {
-  const xhr = new XMLHttpRequest();
-  xhr.open('GET', '.ninrc', false);
-  xhr.send(null);
-
-  return JSON.parse(xhr.responseText);
-}
-
-module.exports = ninrc;
