+++ nin/frontend/app/scripts/utils/ninrc.js
@@ -1,0 +1,5 @@
+const xhr = new XMLHttpRequest();
+xhr.open('GET', '.ninrc', false);
+xhr.send(null);
+
+module.exports = JSON.parse(xhr.responseText);
