+++ nin/frontend/app/styles/common.less
@@ -167,7 +167,6 @@ body.fullscreen .bottom {
 .musiclayer .over {
     overflow: hidden;
     position: relative;
-    top: -50px;
     height: 100%;
 }
 
@@ -568,8 +567,8 @@ svg text.monospaced {
   height: 100px;
 }
 
-demo .overlay {
-    position: absolute; 
+.overlay {
+    position: absolute;
     z-index: 20;
     background: transparent;
 }
