--- nin/frontend/app/views/bottom-panel.html
@@ -1,79 +1,0 @@
-<div class=bottom>
-
-  <div class="fixed-panel">
-
-    <div class="frame-panel">
-      <label>frame</label>
-      <span
-       class=value
-       title="Global frame for demo"
-        >
-        {{ main.demo.getCurrentFrame() }}
-      </span>
-    </div>
-    <div class="beat-bean-panel">
-      <label>BEAN</label>
-      <span
-        class=value
-        title="Global music BEAT number for frame"
-        >
-        {{ bottom.$window.BEAN }}
-      </span>
-      </div>
-  </div>
-
-  <div
-    class=layers-bar-container
-    ng-click="bottom.musicLayerClick($event)"
-    >
-    <div
-      class="marker-line play"
-      ng-hide="hideMarker"
-      style="
-        margin-left: {{ main.currentFrame * bottom.xScale }}px;
-        height: {{ 50 + 30 * layers.length }}px;
-      "
-      >
-      <div
-        class="glow glow-play"
-        ng-class="{'glow-visible': !main.demo.music.paused}"
-        >
-        </div>
-    </div>
-
-    <div
-      class="marker-line loop"
-      ng-show="bottom.loopStart"
-      style="
-        margin-left: {{ bottom.loopStart * bottom.xScale }}px;
-        height: {{ 50 + 30 * layers.length }}px;
-      "
-      >
-      <div class="glow glow-loop">
-        </div>
-    </div>
-
-    <div
-      class="marker-line loop"
-      ng-show="bottom.loopEnd"
-      style="
-        margin-left: {{ bottom.loopEnd * bottom.xScale }}px;
-        height: 50px;
-      "
-      >
-      <div class="glow glow-loop">
-        </div>
-    </div>
-
-    <div
-      waveform
-      duration="{{main.duration}}"
-      class="layer musiclayer"
-      ng-click="bottom.musicLayerClick($event)"
-      style="width: {{ main.duration * bottom.xScale }}px">
-      &nbsp;
-    </div>
-
-  </div>
-
-</div>
