--- nin/frontend/app/views/main.html
@@ -1,23 +1,0 @@
-<link rel="stylesheet" ng-href="styles/{{ main.selectedTheme }}.css">
-
-<div
-  class="disconnect-banner"
-  ng-if="main.disconnected"
-  >
-  Disconnected from the nin backend!
-  Try refreshing, and perhaps restarting <code>nin</code> too.
-</div>
-
-<div menubar menu=main.menu></div>
-
-<div
-  ng-include="'views/top-panel.html'"
-  ng-controller="TopCtrl as top"
-  >
-</div>
-
-<div
-  ng-include="'views/bottom-panel.html'"
-  ng-controller="BottomCtrl as bottom"
-  >
-</div>
