--- nin/frontend/app/views/menubar.html
@@ -1,36 +1,0 @@
-<div class=top-menu>
-
-  <div
-    class=menu-group
-    ng-repeat="menuGroup in menu"
-    ng-mouseover="menuState.focus = menuGroup.name"
-    ng-click="menuState.open = true"
-    ng-class="{selected: menuState.open && menuState.focus == menuGroup.name}"
-    >
-    <div class=menu-item>
-      {{ menuGroup.name }}
-    </div>
-    <div
-      class=sub-menu
-      ng-show="menuState.focus == menuGroup.name && menuState.open"
-      ng-click="menuState.open = false"
-      >
-      <div
-        class=menu-item
-        ng-class="{'separator-item': item.name == '-'}"
-        ng-repeat="item in menuGroup.items"
-        ng-click=item.invoke()
-        >
-        <div ng-show="item.name != '-'">
-          {{ item.name }}
-          <span class=shortcut>{{ item.keybind }}</span>
-        </div>
-        <div
-          class=separator
-          ng-show="item.name == '-'"
-          >
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
