--- nin/frontend/app/views/top-panel.html
@@ -1,25 +1,0 @@
-<div class=top-panel>
-  <div class=node-area-container>
-    <div
-      class=graph-editor
-      graph-editor
-      graph=main.graph
-      >
-    </div>
-  </div>
-
-  <div class="floating-viewer-container">
-    <!-- cramped formatting because of whitespace: pre -->
-    <div
-      class=global-error-overlay
-      ng-repeat="(_, error) in main.globalJSErrors"
-      ><div class="context">⛄ {{ error.context }} ⛄</div>
-<div class="bg">
-<div class="header"><strong>{{ error.stack.split('\n')[0] }}</strong></div>
-<div class="excerpt"
-  ng-bind-html="error | errorExcerpt:fileCache"
-  ></div><div class="stack" ng-bind-html="error | stackFormat"></div>
-</div></div>
-    <demo></demo>
-  </div>
-</div>
