+++ nin/frontend/package.json
@@ -2,9 +2,6 @@
   "name": "nin-frontend",
   "version": "0.0.1",
   "dependencies": {
-    "angular": "1.5.8",
-    "angular-ui-bootstrap": "2.1.0",
-    "angular-ui-layout": "1.4.2",
     "bootstrap": "3.3.7",
     "jquery": "3.1.1",
     "jquery-ui": "1.12.1",
@@ -21,6 +18,7 @@
     "babel-core": "^6.24.0",
     "babel-eslint": "^7.1.1",
     "babel-loader": "^6.4.0",
+    "babel-preset-react": "6.24.1",
     "babel-preset-es2015": "^6.24.0",
     "copy-webpack-plugin": "^4.0.1",
     "css-loader": "^0.27.3",
